lab1 v3

+++ Lab1CPPClass/struct1.cpp
@@ -68,7 +68,7 @@ int main(int argc, char* argv[])
 			Array arr  = Array(asize1);
 
 			int n = 0;
-			for (int i = 0; i < asize1; i++) {
+			for (size_t i = 0; i < asize1; i++) {
 				in >> n;
 				arr.set(i, n);
 			}
